Use named exports internally
Instead of `module.exports = ...`
