+++ packages/resolver/lib/sources/index.ts
@@ -1,18 +1,14 @@
-const EthPMv1Source = require("./ethpm-v1");
-const NPMSource = require("./npm");
-const GlobalNPMSource = require("./globalnpm");
-const FSSource = require("./fs");
-
+import { EthPMv1 } from "./ethpm-v1";
+import { NPM } from "./npm";
+import { GlobalNPM } from "./globalnpm";
+import { FS } from "./fs";
 import { ResolverSource } from "../source";
 
-module.exports = (options: any): ResolverSource[] => {
+export function constructSources(options: any): ResolverSource[] {
   return [
-    new EthPMv1Source(
-      options.working_directory,
-      options.contracts_build_directory
-    ),
-    new NPMSource(options.working_directory),
-    new GlobalNPMSource(),
-    new FSSource(options.working_directory, options.contracts_build_directory)
+    new EthPMv1(options.working_directory),
+    new NPM(options.working_directory),
+    new GlobalNPM(),
+    new FS(options.working_directory, options.contracts_build_directory)
   ];
-};
+}
