+++ packages/resolver/test/npm.ts
@@ -1,7 +1,7 @@
 import assert from "assert";
 
-const NPM = require("../lib/sources/npm");
-const npm = new NPM();
+import { NPM } from "../lib/sources/npm";
+const npm = new NPM("");
 
 describe("npm", function() {
   describe("#require()", function() {
